Refactor yalm_tokenizer_src -> yalm_tokenizer_data
